+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unversioned
+
+### Added
+
+- Option for replacing tab stop character with number of spaces.
+- Button for clearing settings.
+
 ## [0.4.1] - 2025-12-04
 
 ### Fixed